+++ src/main/java/org/betterx/betternether/world/structures/plants/LegacyStructureAnchorTree.java
@@ -192,14 +192,14 @@ public class LegacyStructureAnchorTree implements IStructure {
         double x = Mth.lerp(mix, start.getX(), end.getX());
         double y = Mth.lerp(mix, start.getY(), end.getY());
         double z = Mth.lerp(mix, start.getZ(), end.getZ());
-        return new BlockPos(x, y, z);
+        return new BlockPos((int) x, (int) y, (int) z);
     }
 
     private BlockPos lerpCos(BlockPos start, BlockPos end, int y, double mix) {
         double v = lcos(mix);
         double x = Mth.lerp(v, start.getX(), end.getX());
         double z = Mth.lerp(v, start.getZ(), end.getZ());
-        return new BlockPos(x, y, z);
+        return new BlockPos((int) x, (int) y, (int) z);
     }
 
     private double lcos(double mix) {
@@ -215,7 +215,7 @@ public class LegacyStructureAnchorTree implements IStructure {
             double x = Mth.lerp(delta, start.getX(), end.getX()) + random.nextGaussian() * range;
             double y = Mth.lerp(delta, start.getY(), end.getY());
             double z = Mth.lerp(delta, start.getZ(), end.getZ()) + random.nextGaussian() * range;
-            result.add(new BlockPos(x, y, z));
+            result.add(new BlockPos((int) x, (int) y, (int) z));
         }
         result.add(end);
         return result;
@@ -253,7 +253,11 @@ public class LegacyStructureAnchorTree implements IStructure {
         for (int i = 0; i < count; i++) {
             double x = pos.getX() + Math.sin(startAngle) * radius;
             double z = pos.getZ() + Math.cos(startAngle) * radius;
-            BlockPos end = new BlockPos(x, pos.getY() + height + height * random.nextDouble() * 0.5, z);
+            BlockPos end = new BlockPos(
+                    (int) x,
+                    (int) (pos.getY() + height + height * random.nextDouble() * 0.5),
+                    (int) z
+            );
             List<BlockPos> elem = new ArrayList<>(2);
             elem.add(pos);
             elem.add(end);
